% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_site_prob.R
\name{cast_site_prob}
\alias{cast_site_prob}
\title{Subset cast dataset per probability threshold}
\usage{
cast_site_prob(data, prob = 0.5, site = "missing")
}
\arguments{
\item{data}{Input cast dataset}

\item{prob}{Numeric probability threshold value, from which on (>=) species are subset - e.g. 0.5 (standard)}

\item{site}{EBN (El Bosque Nuevo) or PP (Peru Panguana)}
}
\value{
By probability threshold reduced dataset
}
\description{
Takes the input cast dataset and subsets into dataset containing only those species that were part of at least one species pair occurring with a probability >= a certain threshold value.
}

