% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicheoverlap.R
\name{nicheoverlap}
\alias{nicheoverlap}
\title{Niche overlap analysis}
\usage{
nicheoverlap(data, method = "missing", ra4 = TRUE, iterations = 1,
  txt = FALSE, site = "missing")
}
\arguments{
\item{data}{Input data in cast format}

\item{method}{pianka or czekanowski, niche overlap index you want to use}

\item{ra4}{TRUE (standard) or FALSE, if FALSE no RA4 randomization of the input data}

\item{iterations}{1 (standard), numeric value of how many iterations you want the funtion to run}

\item{txt}{TRUE or FALSE (standard), if TRUE automaticly saves organized txt output files}

\item{site}{EBN (El Bosque Nuevo) or PP (Peru Panguana)}
}
\value{
Temporal niche overlap probabilities for species pairs
}
\description{
This function calculates the temporal niche overlap probabilities for each possible species pair within an animal community.
}
\details{
Further, the processing time of the function is displayed: start / end / duration
}

