% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organized_txt.R
\name{organized_txt}
\alias{organized_txt}
\title{Automaticly save txt output files}
\usage{
organized_txt(data_spp, method, site)
}
\description{
Automaticly creates folders within the current working directory and saves organized txt output files.
}
\details{
This function is part of the "nicheoverlap" function.
}

