% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra4_mod.R
\name{ra4_mod}
\alias{ra4_mod}
\title{RA4 niche overlap randomization algorithm}
\usage{
ra4_mod(speciesData)
}
\description{
Randomizes the monitoring data with applying a modification of the RA4 function from the "EcosimR" package to avoid upcoming error messages.
}
\details{
This function is part of the "nicheoverlap" function.
}

