% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin_txt.R
\name{readin_txt}
\alias{readin_txt}
\title{Read in output files}
\usage{
readin_txt(method, site)
}
\arguments{
\item{method}{pianka or czekanowski}

\item{site}{EBN (El Bosque Nuevo) or PP (Peru Panguana)}
}
\description{
Reads in the automaticly saved txt output files saved by the "organized_txt" function. Compiles the results into one dataframe per method and site.
}

